Move the ktlint config out of the android block, tweak it and add some comments.
I'm at a bit of a loss, if I'm honest, as to how it was working when included within the android config block in the first place. There will be some Gradle and/or Groovy magic going on here that I don't yet understand in terms of delegation.
I've removed the explicit setting of outputToConsole as this defaults to true.
I've also explicitly asked it to colour output as that helps humans scan the console output.