Experimentally remove the explicit contents permissions scope.
Now that I know the docs workflow is working, I want to confirm my understanding (from the GitHub docs) that the entire set of permissions we define here will completely override the entire set of permissions set by default for the org.

i.e. my expected outcome from this workflow in CI is for it to fail.