Reduce test boilerplate repetition for the 'executing normally' set of verifications against the ConnectionReadyWorker.
I've experimented with verifyAll instead of verify(exactly = 1) and that gave me an unexpected test failure, which probably means I don't understand the implications fully.

I also experimented with adding calls to confirmVerified against all the mocks in the cleanUp method implementation, but that then revealed a bunch of mock interactions not being explicitly verified in other tests.
We need to consider whether we should be stricter in this sense, and make sure tests validate all interactions.