Put the ably-java client library behind an interface
The DefaultAbly class already has quite a bit of logic (e.g. retries) and I
intend to add further logic to it. It would be good to be able to unit test
this class. It sounds like we haven’t done this so far because we didn’t want
to try and mock the AblyRealtime class. But I think we need to be able to unit
test DefaultAbly (which I’ve created issue #869 for), so I’m putting the
ably-java library behind a set of AblySdk* interfaces. This will allow us to
provide mock implementations of the Ably client library when testing
DefaultAbly.

(I’ve taken this approach from the asset-tracking-swift codebase.)

One thing worth noticing is the new use of the !! operator when extracting the
`reason` from a connection that’s in a failed state. This looks like it’s
introducing a new unsafe behaviour, but it’s actually just making explicit a
previous assumption, namely that when a connection is in a failed state its
`reason` will be non-null. The previous version of the code - in which the
connection’s `reason` was of platform type ErrorInfo! – would equally have
thrown a NullPointerException if this assumption were violated.