Remove timeout from Publisher.stop
It’s not clear for callers of Publisher.stop what it means if it throws
TimeoutCancellationException, or how they should respond to it. For example, do
they need to try calling it again, or is the publisher now in some terminal
broken state?

(The same questions also apply if the publisher throws ConnectionException;
I’ve created #873 for addressing this.)

So, we remove the timeout, and deprecate the parameter that lets users specify
its duration. (This parameter is now ignored.) We add upgrading instructions to
tell users how they can implement a timeout themselves, if they so wish, and
what they should do if the timeout elapses.

We discussed whether to consider this a breaking change, and the consensus
amongst the team seemed to be that we should consider this timeout a bug that
should have never been implemented in the first place, and hence we can remove
it in a minor release. They wanted to keep the existing timeoutInMilliseconds
parameter though, to avoid breaking compilation for users.

Closes #862.