Add a test helper class for DefaultAbly creation
There’s quite a lot of boilerplate required to create all the mocks needed to
instantiate a DefaultAbly instance, and to then provide any other irrelevant
dependencies that are needed for instantiation. Let’s put it into a class so
that future test cases can use it too.

I’m always on the fence about writing these sort of helpers, because it robs us
of some granularity in how we configure the mocks for each test case – you’ll
often end up mocking things that a particular test case doesn’t need because
the majority of test cases do need it. Or if you want to change the helper you
have to think about the impact on all the test cases. And whilst overall I
think that these helpers do improve test readability, they require the reader
of the tests to have a good understanding of exactly what mocks the helper is
setting up. I think it’s the same class of problems as using other
one-size-fits-all test helpers such as Rails’ fixtures or factory_bot, for
example.

But on the other hand, I’m also becoming convinced that test cases need to be
relatively concise since otherwise, for example, code review becomes a
terrifying task. I’m hoping that by documenting the test helper class in lots
of detail we can perhaps address at least my readability concerns from the
previous paragraph.