Remove AblySdkRealtime extension methods in DefaultAbly
When writing unit tests for DefaultAbly, I want my MockK mocks to only mock the
methods defined in the main AblySdkRealtime interface.

I consider the functionality defined in the extension methods to be part of
DefaultAbly and don’t want to mock them, but MockK will raise an error if these
methods get called on an AblySdkRealtime mock without specifying a mocked
implementation. Hence, we remove them from AblySdkRealtime and make them
methods on DefaultAbly instead.