Use different locations in test data file
Prior to this change, the testing data file contained a series of the same locations (excluding timestamp). On receiving a location update, we stamp the current millisecond-precision timestamp on it.

If during tests, mapbox enhanced updates come through very quickly (<1ms), they are registered on the publisher, but may not be sent to Ably.

This may be due to resolution, or because such a quick update would make the timestamp on subsequent updates the same. As everything else is the same, this would produce the same hashcode for the location update, which is also what we use as the message id for the Ably realtime service. So the message would be rejectedfor reasons of idempotency.

This change therefore makes sure that each data point in the file is different, so we can never have clashing hashcodes.