wait for publisher emissions before performing test assertion
Emitting events on the publishers location channel is done on the same coroutine scope
as, but outside of the worker queue. It's also initiated after we send the location update
to any subscriber over the realtime service.

In a very specific order of events, this could cause the test to fail as the publisher
will have sent the events to the subscriber, but not emitted them onto its location channel.

This change fixes the problem by ensuring that the publisher has emitted at least as many
events as received by the subscriber before allowing test assertions to made.

Fixes #899