Add dependency constraints for androidx.fragment:fragment
It's a transitive dependency which one of our direct dependencies was forcing us to use an older version of.
We were using Fragment 1.4 via com.google.android.gms:play-services-location:21.0.1 (latest) and androidx.appcompat:appcompat:1.6.0 (latest).
The latest version of Fragment is 1.5.5, which is what I've overridden to in this commit.