Add cooling off period for L7 proxy in tests
Debugging indicates that whilst ably-java has sent its shutdown and detach
protocol messages, these are not immediately sent by the underlying transport.

This can mean that in some circumstances, the L7 proxy can begin its shutdown
procedure, yet receive more events in the middle of this, potentially causing
it to crash.

This change enforces a grace period on the proxy for shutdown, to give time
for any latent messages from ably-java's transport to arrive.