Change the ably wrapper disconnect method to always succeed
The only point of failure in the method is the `presence.leave()` call which previously was determining whether the whole `disconnect()` operation succeeds or fails. Now, we try to leave the presence for 30 seconds and if that does not happen (either it hangs or fails with a non-retriable error) we simply carry on the disconnect operations as the presence will be left eventually automatically when we detach the channel.