While removing a trackable set its state to "Offline" and make it final
If we don't make the state final, then there's a chance that some events might trigger an automatic state update and override the trackable state to something else.

Additionally, we remove the state flow for the trackable in order to make it clear for users that the trackable is removed.