Allow a delay whilst checking for published location update in tests
During the test, the location updates are fed to mapbox via an Ably channel.
It may take a brief period for the updates to be processed by mapbox and sent
to the publisher to be published on the trackables channel.

This change introduces a delay whilst checking for the location update
to be received, to make sure that this has time to happen.