Fix race condition in retrying presence enter
This change fixes two issues.

1. A race condition on retrying presence enter. At the moment the work to RetryEnterPresence is
queued by AddTrackableWorker. RetryEnterPresenceWorker relies on the trackable being in the publishers
trackables list. The trackable is added to this list by ConnectionReadyWorker, which is itself queued
a few workers down the connection chan from AddTrackableWorker. The race condition is that in some conditions,
especially where flakey connections are involved, the RetryEnterPresenceWorker can run before ConnectionReadyWorker,
meaning that the trackable will never enter presence, and therefore never come online (RetryEnterPresenceWorker does not
queue a re-attempt if the trackable isn't known).

This change fixes the issue by moving the queuing of the RetryEnterPresence work into ConnectionReadyWorker, which ensures
that the trackable is in the publisher properties prior to execution.

2. Trackables entering a failed state

RetryEnterPresenceWorker transitions trackables into the failed state if a fatal error occurs whilst entering presence.
This can happen if the connection re-connects but the channel is not yet attached.

This change fixes the issue by introducing a listener on the ably channel state, which suspends the coroutine until
the channel notifies us of an online state. Thus ensuring that we don't try to enter presence until it is possible to
do so.