Do not retry presence entry on trackable being removed
Add a check that prevents RetryEnterPresenceWorker from being run if
the trackable has been marked for removal.

There is a test case for this already, but it was a false positive pass - as the trackable
had not been added to the publisher properties, so the worker didn't run anyway. The test case has
been amended to check the correct behaviour.