Ignore state updates once continuation resumed
Errors were being thrown because we were trying to resume the coroutines
multiple times as ably connection states changed.

This change fixes this by only allowing the continuation to be resumed
once. This is not the technically ideal solution - however doing things
"properly" (that is, calling ably.channel.off()) would require a not
insignificant refactor. The proposal is therefore that we do this for now
and raise a separate issue to address the wider issue.