Replace callback failures with failed trackable in workers
Previously, any one of the ConnectionCreatedWorker and ConnectionReadyWorker
would call the trackable callback with a failure result if something went wrong.

The callback is now immediately successful and control passed back to the caller. This change makes it so that these workers will now queue up a FailTrackable
job when something goes wrong.