Refactor presence enter workers
At the moment, ConnectionCreatedWorker handles presence entry flags
among other things, and we have EnterPresenceRetryWorker for retrying.

To try and reduce the number of responsibilities, this change attempts to specialise the workers to only performing one role, as follows:

EnterPresenceSuccessWorker: Essentially the same as the old RetryEnterPresenceSuccessWorker, but renamed for dual purpose use. It sets the presence flags and triggers the publisher to update the trackable states.

A test was added for this worker as the previous class appears not to have had one.

EnterPresenceWorker: takes the initial "did ably.connect succeed and hence we entered presence" value and either queues up a EnterPresenceSuccessWorker or RetryPresenceWorker.

RetryEnterPresenceWorker now calls to EnterPresenceSuccessWorker when presence enter succeeds.

AddTrackable worker triggers EnterPresenceWorker after starting the Ably connections.

ConnectionCreatedWorker no longer sets presence flags.