Refactor presence subscription into its own worker, remove ConnectionCreatedWorker
ConnectionCreatedWorker was doing essentially two things, set the publisher to
connected and subscribe to presence (leaving it to the ConnectionReadyWorker to spawn a retry if required).

This change removes ConnectionCreatedWorker and redistributes its responsibilities.

The job of setting the publisher to connected is now in ConnectionReadyWorker.

The job of doing presence subscription is now in a new SubscribeToPresenceWorker. This worker will try (and retry as necessary) to subscribe to presence, and post a success job to the worker queue on completion.