Dont send resolutions until presence enter succeeds
Resolution sending was moved from ConnectionReadyWorker to AddTrackableWorker as part of the work to make track return instantly, however it requires presence to be successfully entered before it can work - as updatePresenceData returns success immediately if the ably channel doesn't yet exist.

This change fixes this issue by moving the sending of resolutions into EnterPresenceSuccessWorker, which does not get called until we know we're in presence.

It also adds some missing tests in AddTrackableWorker.