Add grace period for publisher online checking in test
In cases where the ably channel enters a suspended state, AAT will
mark the trackable as online again as soon as the channel re-attaches,
even whilst ably-java is re-entering presence under the hood.

This fix provides a grace period in the test, to allow for ably-java
to do the presence reconnect.

We should look at making the trackable offline until presence actually
re-enters at some point.