Allow trackable to enter presence after being removed and re-added
Previously, removing a trackable (when there are multiple) and re-adding would cause that trackable to be stuck in an OFFLINE state - because removing the trackable did not remove it from the removalGuard on completion, meaning that presence related operations would never be attempted.

This change fixes the issue by making sure DisconnectSuccessWorker removes the trackable from the guard when removal is complete.

This will not currently fix the issue in our example apps due to an ably-java isssue, which when fixed will resolve https://github.com/ably/ably-asset-tracking-android/issues/988