Fix service leakage through mapbox
For some reason, when using the `notification` member of the service
directly in the notification provider given to mapbox, the service
is leaked. This is because a synthetic lambda in mapbox appears to keep
the service alive.

This change fixes the issue by providing mapbox with a clone of the
notification, thus avoiding the service leakage.