Remove channel listeners on disconnect
This is possibly a bit of a bug with ably-java. When a channel
is released, it performs a detach but does not wait for the completion
of that detach (and the channel is removed from the map preventing
subsequent messages from coming through).

As part of the detach process, a timer is setup to track the detach.
If the channel has any listeners, those listeners are kept alive via
the GC roots whilst the timer is active.

This change temporarily works around this by dropping any listeners
from the channel whilst waiting for the detach (just for AAT).