Add network connectivity tests assertions on new publisher presence state API
This adds assertions to the subscriber tests for the new
Subscriber.publisherPresenceStateChanges API added in 603908a.

It only adds assertions for the `current` property of the state change. I
thought that adding assertions about the other properties might be quite
complicated, and would be easier with unit tests. I’m keen to hear opinions,
though.

Resolves #1037.