Handle channel detachment logic around RTL3e
A connection dropping out does not detach the channel, so our current implementation based on
channel state does not cover all possible cases - as we may have both a connection dropout as well
as a general channel detach.

This change handles both, by beginning to cache presence messages soon as the connection enters
DISCONNECTED or the channel enters DETACHED, but only starting to process them again once the channel
enters the ATTACHED state.